feat: 게시글 삭제 시 비밀번호 확인 기능 추가
- body의 password값과 DB의 password값 비교를 통해 일치 시 게시글 삭제하는 기능 추가
- 일치하지 않는 경우 401 코드와 삭제 권한 없음 메시지 반환

+++ routes/posts.js
@@ -82,7 +82,7 @@ router.put("/posts/:_postId", async (req, res) => {
                 // 201, 수정 완료 메시지 반환
                 return res.status(201).json({ message: "게시글을 수정하였습니다." })
             } else {
-                return res.status(401).json({ message: "삭제 권한이 없습니다."})
+                return res.status(401).json({ message: "삭제 권한이 없습니다." })
             }
             // 검색 결과가 없는 경우 404, 게시글 조회 실패 메시지 반환
         } else {
@@ -95,6 +95,7 @@ router.put("/posts/:_postId", async (req, res) => {
 })
 
 // 5. 게시글 삭제 
+// -API를 호출할 때 입력된 비밀번호를 비교하여 동일할 때만 글이 삭제되게 하기
 router.delete("/posts/:_postId", async (req, res) => {
     // 파라미터로 들어온 _postId를 필드 타입인 Number로 변환해서 저장하고,
     const _postId = Number(req.params._postId)
@@ -107,11 +108,15 @@ router.delete("/posts/:_postId", async (req, res) => {
             .where("postId").equals(_postId)
         // 검색결과가 있는 경우,
         if (result.length) {
-            // DB에서 삭제 후,
-            await Posts.deleteOne({ postId: _postId })
-            // 200, 삭제 완료 메시지 반환
-            return res.status(200).json({ message: "게시글을 삭제하였습니다." })
-            // 검색 결과가 없는 경우 400, 게시글 조회 실패 메시지 반환
+            if (password === result[0].password) {
+                // DB에서 삭제 후,
+                await Posts.deleteOne({ postId: _postId })
+                // 200, 삭제 완료 메시지 반환
+                return res.status(200).json({ message: "게시글을 삭제하였습니다." })
+                // 검색 결과가 없는 경우 400, 게시글 조회 실패 메시지 반환
+            } else {
+                return res.status(401).json({ message: "삭제 권한이 없습니다." })
+            }
         } else {
             return res.status(404).json({ message: "게시글 조회에 실패하였습니다." })
         }
